requirements.txt
agentql==1.13.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fake-http-header==0.3.5
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-agentql==1.0.1
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-tests==0.3.20
langchain-text-splitters==0.3.9
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.2
langsmith==0.4.14
markdown-it-py==4.0.0
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.100.1
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
playwright==1.54.0
pluggy==1.6.0
propcache==0.3.2
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==4.0.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
syrupy==4.9.1
tenacity==9.1.2
tf-playwright-stealth==1.2.0
tiktoken==0.11.0
tqdm==4.67.1
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
vcrpy==7.0.0
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0